chore(cubestore): Split metastore module to multiple modules
