+++ rust/cubestore/cubestore/src/metastore/mod.rs
@@ -5,21 +5,22 @@ pub mod listener;
 pub mod metastore_fs;
 pub mod multi_index;
 pub mod partition;
+mod rocks_store;
+mod rocks_table;
 pub mod schema;
 pub mod source;
 pub mod table;
 pub mod wal;
 
+pub use rocks_store::*;
+pub use rocks_table::*;
+
 use async_trait::async_trait;
-use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
 use log::info;
-use rocksdb::{
-    DBIterator, Direction, IteratorMode, MergeOperands, Options, ReadOptions, Snapshot, WriteBatch,
-    WriteBatchIterator, DB,
-};
-use serde::{Deserialize, Deserializer, Serialize};
-use std::hash::{Hash, Hasher};
-use std::{collections::hash_map::DefaultHasher, env, io::Cursor, sync::Arc, time};
+use rocksdb::{MergeOperands, Options, DB};
+use serde::{Deserialize, Serialize};
+use std::hash::Hash;
+use std::{env, io::Cursor, sync::Arc, time};
 use tokio::fs;
 use tokio::sync::{oneshot, Notify, RwLock};
 
@@ -43,9 +44,10 @@ use crate::table::{Row, TableValue};
 use crate::util::aborting_join_handle::AbortingJoinHandle;
 use crate::util::time_span::warn_long;
 use crate::util::WorkerLoop;
-use crate::CubeError;
+use crate::{meta_store_table_impl, CubeError};
 use arrow::datatypes::TimeUnit::Microsecond;
 use arrow::datatypes::{DataType, Field};
+use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
 use chrono::{DateTime, Utc};
 use chunks::ChunkRocksTable;
 use core::{fmt, mem};
@@ -76,7 +78,6 @@ use std::sync::Mutex;
 use std::time::{Duration, SystemTime};
 use table::Table;
 use table::{TableRocksIndex, TableRocksTable};
-use tokio::fs::File;
 use tokio::sync::broadcast::Sender;
 use wal::WALRocksTable;
 
@@ -109,9 +110,9 @@ macro_rules! data_frame_from {
                     ],
                     rows.iter().map(|r|
                         crate::table::Row::new(vec![
-                            crate::table::TableValue::Int(r.id as i64),
+                            crate::table::TableValue::Int(r.get_id() as i64),
                             $(
-                                crate::table::TableValue::String(crate::format_table_value!(r.row, $variant, $tt))
+                                crate::table::TableValue::String(crate::format_table_value!(r.get_row(), $variant, $tt))
                             ),+
                         ])
                     ).collect()
@@ -139,81 +140,6 @@ macro_rules! base_rocks_secondary_index {
 
             fn is_unique(&self) -> bool {
                 RocksSecondaryIndex::is_unique(self)
-            }
-        }
-    };
-}
-
-#[macro_export]
-macro_rules! rocks_table_impl {
-    ($table: ty, $rocks_table: ident, $table_id: expr, $indexes: block) => {
-        pub(crate) struct $rocks_table<'a> {
-            db: crate::metastore::DbTableRef<'a>,
-        }
-
-        impl<'a> $rocks_table<'a> {
-            pub fn new(db: crate::metastore::DbTableRef<'a>) -> $rocks_table {
-                $rocks_table { db }
-            }
-        }
-
-        impl<'a> RocksTable for $rocks_table<'a> {
-            type T = $table;
-
-            fn db(&self) -> &DB {
-                self.db.db
-            }
-
-            fn snapshot(&self) -> &rocksdb::Snapshot {
-                self.db.snapshot
-            }
-
-            fn mem_seq(&self) -> &crate::metastore::MemorySequence {
-                &self.db.mem_seq
-            }
-
-            fn table_id(&self) -> TableId {
-                $table_id
-            }
-
-            fn index_id(&self, index_num: IndexId) -> IndexId {
-                if index_num > 99 {
-                    panic!("Too big index id: {}", index_num);
-                }
-                $table_id as IndexId + index_num
-            }
-
-            fn deserialize_row<'de, D>(
-                &self,
-                deserializer: D,
-            ) -> Result<$table, <D as Deserializer<'de>>::Error>
-            where
-                D: Deserializer<'de>,
-            {
-                <$table>::deserialize(deserializer)
-            }
-
-            fn indexes() -> Vec<Box<dyn BaseRocksSecondaryIndex<$table>>> {
-                $indexes
-            }
-
-            fn update_event(
-                &self,
-                old_row: IdRow<Self::T>,
-                new_row: IdRow<Self::T>,
-            ) -> MetaStoreEvent {
-                paste::expr! { MetaStoreEvent::[<Update $table>](old_row, new_row) }
-            }
-
-            fn delete_event(&self, row: IdRow<Self::T>) -> MetaStoreEvent {
-                paste::expr! { MetaStoreEvent::[<Delete $table>](row) }
-            }
-        }
-
-        impl<'a> core::fmt::Debug for $rocks_table<'a> {
-            fn fmt(&self, f: &mut core::fmt::Formatter) -> core::fmt::Result {
-                f.write_fmt(format_args!("{}", stringify!($rocks_table)))?;
-                Ok(())
             }
         }
     };
@@ -444,21 +370,19 @@ impl ColumnType {
 impl From<&Column> for parquet::schema::types::Type {
     fn from(column: &Column) -> Self {
         match column.get_column_type() {
-            crate::metastore::ColumnType::String => {
+            ColumnType::String => {
                 types::Type::primitive_type_builder(&column.get_name(), Type::BYTE_ARRAY)
                     .with_converted_type(ConvertedType::UTF8)
                     .with_repetition(Repetition::OPTIONAL)
                     .build()
                     .unwrap()
             }
-            crate::metastore::ColumnType::Int => {
-                types::Type::primitive_type_builder(&column.get_name(), Type::INT64)
-                    .with_converted_type(ConvertedType::INT_64)
-                    .with_repetition(Repetition::OPTIONAL)
-                    .build()
-                    .unwrap()
-            }
-            crate::metastore::ColumnType::Decimal { precision, .. } => {
+            ColumnType::Int => types::Type::primitive_type_builder(&column.get_name(), Type::INT64)
+                .with_converted_type(ConvertedType::INT_64)
+                .with_repetition(Repetition::OPTIONAL)
+                .build()
+                .unwrap(),
+            ColumnType::Decimal { precision, .. } => {
                 types::Type::primitive_type_builder(&column.get_name(), Type::INT64)
                     .with_converted_type(ConvertedType::DECIMAL)
                     .with_precision(*precision)
@@ -467,14 +391,14 @@ impl From<&Column> for parquet::schema::types::Type {
                     .build()
                     .unwrap()
             }
-            crate::metastore::ColumnType::Bytes | ColumnType::HyperLogLog(_) => {
+            ColumnType::Bytes | ColumnType::HyperLogLog(_) => {
                 types::Type::primitive_type_builder(&column.get_name(), Type::BYTE_ARRAY)
                     .with_converted_type(ConvertedType::NONE)
                     .with_repetition(Repetition::OPTIONAL)
                     .build()
                     .unwrap()
             }
-            crate::metastore::ColumnType::Timestamp => {
+            ColumnType::Timestamp => {
                 types::Type::primitive_type_builder(&column.get_name(), Type::INT64)
                     //TODO MICROS?
                     .with_converted_type(ConvertedType::TIMESTAMP_MICROS)
@@ -482,7 +406,7 @@ impl From<&Column> for parquet::schema::types::Type {
                     .build()
                     .unwrap()
             }
-            crate::metastore::ColumnType::Boolean => {
+            ColumnType::Boolean => {
                 types::Type::primitive_type_builder(&column.get_name(), Type::BOOLEAN)
                     .with_repetition(Repetition::OPTIONAL)
                     .build()
@@ -707,128 +631,6 @@ pub struct WAL {
     table_id: u64,
     row_count: u64,
     uploaded: bool,
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug, Eq, PartialEq)]
-pub struct IdRow<T: Clone> {
-    id: u64,
-    row: T,
-}
-
-impl<T: Clone> IdRow<T> {
-    pub fn new(id: u64, row: T) -> IdRow<T> {
-        IdRow { id, row }
-    }
-    pub fn get_id(&self) -> u64 {
-        self.id
-    }
-
-    pub fn get_row(&self) -> &T {
-        &self.row
-    }
-
-    pub fn into_row(self) -> T {
-        self.row
-    }
-}
-
-struct KeyVal {
-    key: Vec<u8>,
-    val: Vec<u8>,
-}
-
-struct BatchPipe<'a> {
-    db: &'a DB,
-    write_batch: WriteBatch,
-    events: Vec<MetaStoreEvent>,
-    invalidate_tables_cache: bool,
-}
-
-impl<'a> BatchPipe<'a> {
-    fn new(db: &'a DB) -> BatchPipe<'a> {
-        BatchPipe {
-            db,
-            write_batch: WriteBatch::default(),
-            events: Vec::new(),
-            invalidate_tables_cache: false,
-        }
-    }
-
-    fn batch(&mut self) -> &mut WriteBatch {
-        &mut self.write_batch
-    }
-
-    fn add_event(&mut self, event: MetaStoreEvent) {
-        self.events.push(event);
-    }
-
-    fn batch_write_rows(self) -> Result<Vec<MetaStoreEvent>, CubeError> {
-        let db = self.db;
-        db.write(self.write_batch)?;
-        Ok(self.events)
-    }
-
-    fn invalidate_tables_cache(&mut self) {
-        self.invalidate_tables_cache = true;
-    }
-}
-
-#[derive(Clone)]
-pub struct DbTableRef<'a> {
-    pub db: &'a DB,
-    pub snapshot: &'a Snapshot<'a>,
-    pub mem_seq: MemorySequence,
-}
-
-#[async_trait]
-pub trait MetaStoreTable: Send + Sync {
-    type T: Serialize + Clone + Debug + 'static;
-
-    async fn all_rows(&self) -> Result<Vec<IdRow<Self::T>>, CubeError>;
-
-    async fn row_by_id_or_not_found(&self, id: u64) -> Result<IdRow<Self::T>, CubeError>;
-
-    async fn delete(&self, id: u64) -> Result<IdRow<Self::T>, CubeError>;
-}
-
-#[macro_export]
-macro_rules! meta_store_table_impl {
-    ($name: ident, $table: ty, $rocks_table: ident) => {
-        pub struct $name {
-            rocks_meta_store: RocksMetaStore,
-        }
-
-        impl $name {
-            fn table<'a>(db: DbTableRef<'a>) -> $rocks_table<'a> {
-                <$rocks_table>::new(db)
-            }
-        }
-
-        #[async_trait]
-        impl MetaStoreTable for $name {
-            type T = $table;
-
-            async fn all_rows(&self) -> Result<Vec<IdRow<Self::T>>, CubeError> {
-                self.rocks_meta_store
-                    .read_operation_out_of_queue(move |db_ref| Ok(Self::table(db_ref).all_rows()?))
-                    .await
-            }
-
-            async fn row_by_id_or_not_found(&self, id: u64) -> Result<IdRow<Self::T>, CubeError> {
-                self.rocks_meta_store
-                    .read_operation(move |db_ref| Ok(Self::table(db_ref).get_row_or_not_found(id)?))
-                    .await
-            }
-
-            async fn delete(&self, id: u64) -> Result<IdRow<Self::T>, CubeError> {
-                self.rocks_meta_store
-                    .write_operation(
-                        move |db_ref, batch| Ok(Self::table(db_ref).delete(id, batch)?),
-                    )
-                    .await
-            }
-        }
-    };
 }
 
 meta_store_table_impl!(SchemaMetaStoreTable, Schema, SchemaRocksTable);
@@ -1163,131 +965,6 @@ pub enum MetaStoreEvent {
     DeleteMultiPartition(IdRow<MultiPartition>),
 }
 
-type SecondaryKey = Vec<u8>;
-type IndexId = u32;
-
-#[derive(Debug, Clone, Serialize, Deserialize, Hash, Eq, PartialEq)]
-pub enum RowKey {
-    Table(TableId, u64),
-    Sequence(TableId),
-    SecondaryIndex(IndexId, SecondaryKey, u64),
-    SecondaryIndexInfo { index_id: IndexId },
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Hash, Eq, PartialEq)]
-pub struct SecondaryIndexInfo {
-    version: u32,
-}
-
-pub fn get_fixed_prefix() -> usize {
-    13
-}
-
-impl RowKey {
-    fn from_bytes(bytes: &[u8]) -> RowKey {
-        let mut reader = Cursor::new(bytes);
-        match reader.read_u8().unwrap() {
-            1 => RowKey::Table(TableId::from(reader.read_u32::<BigEndian>().unwrap()), {
-                // skip zero for fixed key padding
-                reader.read_u64::<BigEndian>().unwrap();
-                reader.read_u64::<BigEndian>().unwrap()
-            }),
-            2 => RowKey::Sequence(TableId::from(reader.read_u32::<BigEndian>().unwrap())),
-            3 => {
-                let table_id = IndexId::from(reader.read_u32::<BigEndian>().unwrap());
-                let mut secondary_key: SecondaryKey = SecondaryKey::new();
-                let sc_length = bytes.len() - 13;
-                for _i in 0..sc_length {
-                    secondary_key.push(reader.read_u8().unwrap());
-                }
-                let row_id = reader.read_u64::<BigEndian>().unwrap();
-
-                RowKey::SecondaryIndex(table_id, secondary_key, row_id)
-            }
-            4 => {
-                let index_id = IndexId::from(reader.read_u32::<BigEndian>().unwrap());
-
-                RowKey::SecondaryIndexInfo { index_id }
-            }
-            v => panic!("Unknown key prefix: {}", v),
-        }
-    }
-
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut wtr = vec![];
-        match self {
-            RowKey::Table(table_id, row_id) => {
-                wtr.write_u8(1).unwrap();
-                wtr.write_u32::<BigEndian>(*table_id as u32).unwrap();
-                wtr.write_u64::<BigEndian>(0).unwrap();
-                wtr.write_u64::<BigEndian>(row_id.clone()).unwrap();
-            }
-            RowKey::Sequence(table_id) => {
-                wtr.write_u8(2).unwrap();
-                wtr.write_u32::<BigEndian>(*table_id as u32).unwrap();
-            }
-            RowKey::SecondaryIndex(index_id, secondary_key, row_id) => {
-                wtr.write_u8(3).unwrap();
-                wtr.write_u32::<BigEndian>(*index_id as IndexId).unwrap();
-                for &n in secondary_key {
-                    wtr.write_u8(n).unwrap();
-                }
-                wtr.write_u64::<BigEndian>(row_id.clone()).unwrap();
-            }
-            RowKey::SecondaryIndexInfo { index_id } => {
-                wtr.write_u8(4).unwrap();
-                wtr.write_u32::<BigEndian>(*index_id as IndexId).unwrap();
-            }
-        }
-        wtr
-    }
-}
-
-macro_rules! enum_from_primitive_impl {
-    ($name:ident, $( $variant:ident )*) => {
-        impl From<u32> for $name {
-            fn from(n: u32) -> Self {
-                $( if n == $name::$variant as u32 {
-                    $name::$variant
-                } else )* {
-                    panic!("Unknown {}: {}", stringify!($name), n);
-                }
-            }
-        }
-    };
-}
-
-macro_rules! enum_from_primitive {
-    (
-        $( #[$enum_attr:meta] )*
-        pub enum $name:ident {
-            $( $( $( #[$variant_attr:meta] )* $variant:ident ),+ = $discriminator:expr ),*
-        }
-    ) => {
-        $( #[$enum_attr] )*
-        pub enum $name {
-            $( $( $( #[$variant_attr] )* $variant ),+ = $discriminator ),*
-        }
-        enum_from_primitive_impl! { $name, $( $( $variant )+ )* }
-    };
-}
-
-enum_from_primitive! {
-    #[derive(Copy, Clone, Eq, PartialEq, Debug, Serialize, Deserialize, Hash)]
-    pub enum TableId {
-        Schemas = 0x0100,
-        Tables = 0x0200,
-        Indexes = 0x0300,
-        Partitions = 0x0400,
-        Chunks = 0x0500,
-        WALs = 0x0600,
-        Jobs = 0x0700,
-        Sources = 0x0800,
-        MultiIndexes = 0x0900,
-        MultiPartitions = 0x0A00
-    }
-}
-
 fn check_indexes_for_all_tables<'a>(table_ref: DbTableRef<'a>) -> Result<(), CubeError> {
     SchemaRocksTable::new(table_ref.clone()).check_indexes()?;
     TableRocksTable::new(table_ref.clone()).check_indexes()?;
@@ -1300,21 +977,6 @@ fn check_indexes_for_all_tables<'a>(table_ref: DbTableRef<'a>) -> Result<(), Cub
     MultiIndexRocksTable::new(table_ref.clone()).check_indexes()?;
     MultiPartitionRocksTable::new(table_ref.clone()).check_indexes()?;
     Ok(())
-}
-
-#[derive(Clone)]
-pub struct MemorySequence {
-    seq_store: Arc<Mutex<HashMap<TableId, u64>>>,
-}
-
-impl MemorySequence {
-    pub fn next_seq(&self, table_id: TableId, snapshot_value: u64) -> Result<u64, CubeError> {
-        let mut store = self.seq_store.lock()?;
-        let mut current = *store.entry(table_id).or_insert(snapshot_value);
-        current += 1;
-        store.insert(table_id, current);
-        Ok(current)
-    }
 }
 
 #[derive(Clone)]
@@ -1335,756 +997,6 @@ pub struct RocksMetaStore {
         Box<dyn FnOnce() -> Result<(), CubeError> + Send + Sync + 'static>,
     >,
     _rw_loop_join_handle: Arc<AbortingJoinHandle<()>>,
-}
-
-trait BaseRocksSecondaryIndex<T>: Debug {
-    fn index_key_by(&self, row: &T) -> Vec<u8>;
-
-    fn get_id(&self) -> u32;
-
-    fn key_hash(&self, row: &T) -> u64 {
-        let key_bytes = self.index_key_by(row);
-        self.hash_bytes(&key_bytes)
-    }
-
-    fn hash_bytes(&self, key_bytes: &Vec<u8>) -> u64 {
-        let mut hasher = DefaultHasher::new();
-        key_bytes.hash(&mut hasher);
-        hasher.finish()
-    }
-
-    fn is_unique(&self) -> bool;
-
-    fn version(&self) -> u32;
-}
-
-trait RocksSecondaryIndex<T, K: Hash>: BaseRocksSecondaryIndex<T> {
-    fn typed_key_by(&self, row: &T) -> K;
-
-    fn key_to_bytes(&self, key: &K) -> Vec<u8>;
-
-    fn typed_key_hash(&self, row_key: &K) -> u64 {
-        let key_bytes = self.key_to_bytes(row_key);
-        self.hash_bytes(&key_bytes)
-    }
-
-    fn index_key_by(&self, row: &T) -> Vec<u8> {
-        self.key_to_bytes(&self.typed_key_by(row))
-    }
-
-    fn get_id(&self) -> u32;
-
-    fn is_unique(&self) -> bool;
-
-    fn version(&self) -> u32;
-}
-
-impl<T, I> BaseRocksSecondaryIndex<T> for I
-where
-    I: RocksSecondaryIndex<T, String>,
-{
-    fn index_key_by(&self, row: &T) -> Vec<u8> {
-        self.key_to_bytes(&self.typed_key_by(row))
-    }
-
-    fn get_id(&self) -> u32 {
-        RocksSecondaryIndex::get_id(self)
-    }
-
-    fn is_unique(&self) -> bool {
-        RocksSecondaryIndex::is_unique(self)
-    }
-
-    fn version(&self) -> u32 {
-        RocksSecondaryIndex::version(self)
-    }
-}
-
-struct TableScanIter<'a, RT: RocksTable + ?Sized> {
-    table_id: TableId,
-    table: &'a RT,
-    iter: DBIterator<'a>,
-}
-
-impl<'a, RT: RocksTable<T = T> + ?Sized, T> Iterator for TableScanIter<'a, RT>
-where
-    T: Serialize + Clone + Debug + Send,
-{
-    type Item = Result<IdRow<T>, CubeError>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let option = self.iter.next();
-        if let Some((key, value)) = option {
-            if let RowKey::Table(table_id, row_id) = RowKey::from_bytes(&key) {
-                if table_id != self.table_id {
-                    return None;
-                }
-                Some(self.table.deserialize_id_row(row_id, &value))
-            } else {
-                None
-            }
-        } else {
-            None
-        }
-    }
-}
-
-struct IndexScanIter<'a, RT: RocksTable + ?Sized> {
-    table: &'a RT,
-    secondary_key_val: Vec<u8>,
-    secondary_key_hash: Vec<u8>,
-    iter: DBIterator<'a>,
-}
-
-impl<'a, RT: RocksTable<T = T> + ?Sized, T> Iterator for IndexScanIter<'a, RT>
-where
-    T: Serialize + Clone + Debug + Send,
-{
-    type Item = Result<IdRow<T>, CubeError>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        loop {
-            let option = self.iter.next();
-            if let Some((key, value)) = option {
-                if let RowKey::SecondaryIndex(_, secondary_index_hash, row_id) =
-                    RowKey::from_bytes(&key)
-                {
-                    if &secondary_index_hash != self.secondary_key_hash.as_slice() {
-                        return None;
-                    }
-
-                    if self.secondary_key_val.as_slice() != value.as_ref() {
-                        continue;
-                    }
-
-                    return Some(self.table.get_row_or_not_found(row_id));
-                };
-            } else {
-                return None;
-            }
-        }
-    }
-}
-
-trait RocksTable: Debug + Send + Sync {
-    type T: Serialize + Clone + Debug + Send;
-    fn delete_event(&self, row: IdRow<Self::T>) -> MetaStoreEvent;
-    fn update_event(&self, old_row: IdRow<Self::T>, new_row: IdRow<Self::T>) -> MetaStoreEvent;
-    fn db(&self) -> &DB;
-    fn snapshot(&self) -> &Snapshot;
-    fn mem_seq(&self) -> &MemorySequence;
-    fn index_id(&self, index_num: IndexId) -> IndexId;
-    fn table_id(&self) -> TableId;
-    fn deserialize_row<'de, D>(&self, deserializer: D) -> Result<Self::T, D::Error>
-    where
-        D: Deserializer<'de>;
-    fn indexes() -> Vec<Box<dyn BaseRocksSecondaryIndex<Self::T>>>;
-
-    fn insert(
-        &self,
-        row: Self::T,
-        batch_pipe: &mut BatchPipe,
-    ) -> Result<IdRow<Self::T>, CubeError> {
-        let mut ser = flexbuffers::FlexbufferSerializer::new();
-        row.serialize(&mut ser).unwrap();
-        let serialized_row = ser.take_buffer();
-
-        for index in Self::indexes().iter() {
-            let hash = index.key_hash(&row);
-            let index_val = index.index_key_by(&row);
-            let existing_keys =
-                self.get_row_from_index(index.get_id(), &index_val, &hash.to_be_bytes().to_vec())?;
-            if index.is_unique() && existing_keys.len() > 0 {
-                return Err(CubeError::user(
-                    format!(
-                        "Unique constraint violation: row {:?} has a key that already exists in {:?} index",
-                        &row,
-                        index
-                    )
-                ));
-            }
-        }
-
-        let (row_id, inserted_row) = self.insert_row(serialized_row)?;
-        batch_pipe.add_event(MetaStoreEvent::Insert(self.table_id(), row_id));
-        if self.snapshot().get(&inserted_row.key)?.is_some() {
-            return Err(CubeError::internal(format!("Primary key constraint violation. Primary key already exists for a row id {}: {:?}", row_id, &row)));
-        }
-        batch_pipe.batch().put(inserted_row.key, inserted_row.val);
-
-        let index_row = self.insert_index_row(&row, row_id)?;
-        for to_insert in index_row {
-            if self.snapshot().get(&to_insert.key)?.is_some() {
-                return Err(CubeError::internal(format!("Primary key constraint violation in secondary index. Primary key already exists for a row id {}: {:?}", row_id, &row)));
-            }
-            batch_pipe.batch().put(to_insert.key, to_insert.val);
-        }
-
-        Ok(IdRow::new(row_id, row))
-    }
-
-    fn check_indexes(&self) -> Result<(), CubeError> {
-        let snapshot = self.snapshot();
-        for index in Self::indexes().into_iter() {
-            let index_info = snapshot.get(
-                &RowKey::SecondaryIndexInfo {
-                    index_id: self.index_id(index.get_id()),
-                }
-                .to_bytes(),
-            )?;
-            if let Some(index_info) = index_info {
-                let index_info = self.deserialize_index_info(index_info.as_slice())?;
-                if index_info.version != index.version() {
-                    self.rebuild_index(&index)?;
-                }
-            } else {
-                self.rebuild_index(&index)?;
-            }
-        }
-        Ok(())
-    }
-
-    fn deserialize_index_info(&self, buffer: &[u8]) -> Result<SecondaryIndexInfo, CubeError> {
-        let r = flexbuffers::Reader::get_root(&buffer).unwrap();
-        let row = SecondaryIndexInfo::deserialize(r)?;
-        Ok(row)
-    }
-
-    fn serialize_index_info(&self, index_info: SecondaryIndexInfo) -> Result<Vec<u8>, CubeError> {
-        let mut ser = flexbuffers::FlexbufferSerializer::new();
-        index_info.serialize(&mut ser).unwrap();
-        let serialized_row = ser.take_buffer();
-        Ok(serialized_row)
-    }
-
-    fn rebuild_index(
-        &self,
-        index: &Box<dyn BaseRocksSecondaryIndex<Self::T>>,
-    ) -> Result<(), CubeError> {
-        let time = SystemTime::now();
-        let mut batch = WriteBatch::default();
-        self.delete_all_rows_from_index(index.get_id(), &mut batch)?;
-
-        let all_rows = self.scan_all_rows()?;
-
-        let mut log_shown = false;
-
-        for row in all_rows {
-            if !log_shown {
-                log::info!(
-                    "Rebuilding metastore index {:?} for table {:?}",
-                    index,
-                    self
-                );
-                log_shown = true;
-            }
-            let row = row?;
-            let index_row = self.index_key_val(row.get_row(), row.get_id(), index);
-            batch.put(index_row.key, index_row.val);
-        }
-        batch.put(
-            &RowKey::SecondaryIndexInfo {
-                index_id: self.index_id(index.get_id()),
-            }
-            .to_bytes(),
-            self.serialize_index_info(SecondaryIndexInfo {
-                version: index.version(),
-            })?
-            .as_slice(),
-        );
-        self.db().write(batch)?;
-        if log_shown {
-            log::info!(
-                "Rebuilding metastore index {:?} for table {:?} complete ({:?})",
-                index,
-                self,
-                time.elapsed()?
-            );
-        }
-        Ok(())
-    }
-
-    fn get_row_ids_by_index<K: Debug>(
-        &self,
-        row_key: &K,
-        secondary_index: &impl RocksSecondaryIndex<Self::T, K>,
-    ) -> Result<Vec<u64>, CubeError>
-    where
-        K: Hash,
-    {
-        let hash = secondary_index.typed_key_hash(&row_key);
-        let index_val = secondary_index.key_to_bytes(&row_key);
-        let existing_keys = self.get_row_from_index(
-            RocksSecondaryIndex::get_id(secondary_index),
-            &index_val,
-            &hash.to_be_bytes().to_vec(),
-        )?;
-
-        Ok(existing_keys)
-    }
-
-    fn get_rows_by_index<K: Debug>(
-        &self,
-        row_key: &K,
-        secondary_index: &impl RocksSecondaryIndex<Self::T, K>,
-    ) -> Result<Vec<IdRow<Self::T>>, CubeError>
-    where
-        K: Hash,
-    {
-        let row_ids = self.get_row_ids_by_index(row_key, secondary_index)?;
-
-        let mut res = Vec::new();
-
-        for id in row_ids {
-            if let Some(row) = self.get_row(id)? {
-                res.push(row);
-            } else {
-                let index = Self::indexes()
-                    .into_iter()
-                    .find(|i| i.get_id() == BaseRocksSecondaryIndex::get_id(secondary_index))
-                    .unwrap();
-                self.rebuild_index(&index)?;
-                return Err(CubeError::internal(format!(
-                    "Row exists in secondary index however missing in {:?} table: {}. Repairing index.",
-                    self, id
-                )));
-            }
-        }
-
-        if RocksSecondaryIndex::is_unique(secondary_index) && res.len() > 1 {
-            return Err(CubeError::internal(format!(
-                "Unique index expected but found multiple values in {:?} table: {:?}",
-                self, res
-            )));
-        }
-
-        Ok(res)
-    }
-
-    fn get_single_row_by_index<K: Debug>(
-        &self,
-        row_key: &K,
-        secondary_index: &impl RocksSecondaryIndex<Self::T, K>,
-    ) -> Result<IdRow<Self::T>, CubeError>
-    where
-        K: Hash,
-    {
-        let rows = self.get_rows_by_index(row_key, secondary_index)?;
-        Ok(rows.into_iter().nth(0).ok_or(CubeError::internal(format!(
-            "One value expected in {:?} for {:?} but nothing found",
-            self, row_key
-        )))?)
-    }
-
-    fn update_with_fn(
-        &self,
-        row_id: u64,
-        update_fn: impl FnOnce(&Self::T) -> Self::T,
-        batch_pipe: &mut BatchPipe,
-    ) -> Result<IdRow<Self::T>, CubeError> {
-        let row = self.get_row_or_not_found(row_id)?;
-        let new_row = update_fn(&row.get_row());
-        self.update(row_id, new_row, &row.get_row(), batch_pipe)
-    }
-
-    fn update_with_res_fn(
-        &self,
-        row_id: u64,
-        update_fn: impl FnOnce(&Self::T) -> Result<Self::T, CubeError>,
-        batch_pipe: &mut BatchPipe,
-    ) -> Result<IdRow<Self::T>, CubeError> {
-        let row = self.get_row_or_not_found(row_id)?;
-        let new_row = update_fn(&row.get_row())?;
-        self.update(row_id, new_row, &row.get_row(), batch_pipe)
-    }
-
-    fn update(
-        &self,
-        row_id: u64,
-        new_row: Self::T,
-        old_row: &Self::T,
-        batch_pipe: &mut BatchPipe,
-    ) -> Result<IdRow<Self::T>, CubeError> {
-        let deleted_row = self.delete_index_row(&old_row, row_id)?;
-        for row in deleted_row {
-            batch_pipe.batch().delete(row.key);
-        }
-
-        let mut ser = flexbuffers::FlexbufferSerializer::new();
-        new_row.serialize(&mut ser).unwrap();
-        let serialized_row = ser.take_buffer();
-
-        let updated_row = self.update_row(row_id, serialized_row)?;
-        batch_pipe.add_event(MetaStoreEvent::Update(self.table_id(), row_id));
-        batch_pipe.add_event(self.update_event(
-            IdRow::new(row_id, old_row.clone()),
-            IdRow::new(row_id, new_row.clone()),
-        ));
-        batch_pipe.batch().put(updated_row.key, updated_row.val);
-
-        let index_row = self.insert_index_row(&new_row, row_id)?;
-        for row in index_row {
-            batch_pipe.batch().put(row.key, row.val);
-        }
-        Ok(IdRow::new(row_id, new_row))
-    }
-
-    fn delete(&self, row_id: u64, batch_pipe: &mut BatchPipe) -> Result<IdRow<Self::T>, CubeError> {
-        let row = self.get_row_or_not_found(row_id)?;
-        let deleted_row = self.delete_index_row(row.get_row(), row_id)?;
-        batch_pipe.add_event(MetaStoreEvent::Delete(self.table_id(), row_id));
-        batch_pipe.add_event(self.delete_event(row.clone()));
-        for row in deleted_row {
-            batch_pipe.batch().delete(row.key);
-        }
-
-        batch_pipe.batch().delete(self.delete_row(row_id)?.key);
-
-        Ok(row)
-    }
-
-    fn next_table_seq(&self) -> Result<u64, CubeError> {
-        let ref db = self.db();
-        let seq_key = RowKey::Sequence(self.table_id());
-        let before_merge = self
-            .snapshot()
-            .get(seq_key.to_bytes())?
-            .map(|v| Cursor::new(v).read_u64::<BigEndian>().unwrap());
-
-        // TODO revert back merge operator if locking works
-        let next_seq = self
-            .mem_seq()
-            .next_seq(self.table_id(), before_merge.unwrap_or(0))?;
-
-        let mut to_write = vec![];
-        to_write.write_u64::<BigEndian>(next_seq)?;
-        db.put(seq_key.to_bytes(), to_write)?;
-
-        Ok(next_seq)
-    }
-
-    fn insert_row(&self, row: Vec<u8>) -> Result<(u64, KeyVal), CubeError> {
-        let next_seq = self.next_table_seq()?;
-        let t = RowKey::Table(self.table_id(), next_seq);
-        let res = KeyVal {
-            key: t.to_bytes(),
-            val: row,
-        };
-        Ok((next_seq, res))
-    }
-
-    fn update_row(&self, row_id: u64, row: Vec<u8>) -> Result<KeyVal, CubeError> {
-        let t = RowKey::Table(self.table_id(), row_id);
-        let res = KeyVal {
-            key: t.to_bytes(),
-            val: row,
-        };
-        Ok(res)
-    }
-
-    fn delete_row(&self, row_id: u64) -> Result<KeyVal, CubeError> {
-        let t = RowKey::Table(self.table_id(), row_id);
-        let res = KeyVal {
-            key: t.to_bytes(),
-            val: vec![],
-        };
-        Ok(res)
-    }
-
-    fn get_row_or_not_found(&self, row_id: u64) -> Result<IdRow<Self::T>, CubeError> {
-        self.get_row(row_id)?.ok_or(CubeError::user(format!(
-            "Row with id {} is not found for {:?}",
-            row_id, self
-        )))
-    }
-
-    fn get_row(&self, row_id: u64) -> Result<Option<IdRow<Self::T>>, CubeError> {
-        let ref db = self.snapshot();
-        let res = db.get(RowKey::Table(self.table_id(), row_id).to_bytes())?;
-
-        if let Some(buffer) = res {
-            let row = self.deserialize_id_row(row_id, buffer.as_slice())?;
-            return Ok(Some(row));
-        }
-
-        Ok(None)
-    }
-
-    fn deserialize_id_row(&self, row_id: u64, buffer: &[u8]) -> Result<IdRow<Self::T>, CubeError> {
-        let r = flexbuffers::Reader::get_root(&buffer).unwrap();
-        let row = self.deserialize_row(r)?;
-        return Ok(IdRow::new(row_id, row));
-    }
-
-    fn insert_index_row(&self, row: &Self::T, row_id: u64) -> Result<Vec<KeyVal>, CubeError> {
-        let mut res = Vec::new();
-        for index in Self::indexes().iter() {
-            res.push(self.index_key_val(row, row_id, index));
-        }
-        Ok(res)
-    }
-
-    fn index_key_val(
-        &self,
-        row: &Self::T,
-        row_id: u64,
-        index: &Box<dyn BaseRocksSecondaryIndex<Self::T>>,
-    ) -> KeyVal {
-        let hash = index.key_hash(row);
-        let index_val = index.index_key_by(row);
-        let key = RowKey::SecondaryIndex(
-            self.index_id(index.get_id()),
-            hash.to_be_bytes().to_vec(),
-            row_id,
-        );
-        KeyVal {
-            key: key.to_bytes(),
-            val: index_val,
-        }
-    }
-
-    fn delete_index_row(&self, row: &Self::T, row_id: u64) -> Result<Vec<KeyVal>, CubeError> {
-        let mut res = Vec::new();
-        for index in Self::indexes().iter() {
-            let hash = index.key_hash(&row);
-            let key = RowKey::SecondaryIndex(
-                self.index_id(index.get_id()),
-                hash.to_be_bytes().to_vec(),
-                row_id,
-            );
-            res.push(KeyVal {
-                key: key.to_bytes(),
-                val: vec![],
-            });
-        }
-
-        Ok(res)
-    }
-
-    fn get_row_from_index(
-        &self,
-        secondary_id: u32,
-        secondary_key_val: &Vec<u8>,
-        secondary_key_hash: &Vec<u8>,
-    ) -> Result<Vec<u64>, CubeError> {
-        let ref db = self.snapshot();
-        let key_len = secondary_key_hash.len();
-        let key_min =
-            RowKey::SecondaryIndex(self.index_id(secondary_id), secondary_key_hash.clone(), 0);
-
-        let mut res: Vec<u64> = Vec::new();
-
-        let mut opts = ReadOptions::default();
-        opts.set_prefix_same_as_start(true);
-        let iter = db.iterator_opt(
-            IteratorMode::From(&key_min.to_bytes()[0..(key_len + 5)], Direction::Forward),
-            opts,
-        );
-
-        for (key, value) in iter {
-            if let RowKey::SecondaryIndex(_, secondary_index_hash, row_id) =
-                RowKey::from_bytes(&key)
-            {
-                if !secondary_index_hash
-                    .iter()
-                    .zip(secondary_key_hash)
-                    .all(|(a, b)| a == b)
-                {
-                    break;
-                }
-
-                if secondary_key_val.len() != value.len()
-                    || !value.iter().zip(secondary_key_val).all(|(a, b)| a == b)
-                {
-                    continue;
-                }
-                res.push(row_id);
-            };
-        }
-        Ok(res)
-    }
-
-    fn delete_all_rows_from_index(
-        &self,
-        secondary_id: u32,
-        batch: &mut WriteBatch,
-    ) -> Result<(), CubeError> {
-        let ref db = self.snapshot();
-        let zero_vec = vec![0 as u8; 8];
-        let key_len = zero_vec.len();
-        let key_min = RowKey::SecondaryIndex(self.index_id(secondary_id), zero_vec.clone(), 0);
-
-        let iter = db.iterator(IteratorMode::From(
-            &key_min.to_bytes()[0..(key_len + 5)],
-            Direction::Forward,
-        ));
-
-        for (key, _) in iter {
-            let row_key = RowKey::from_bytes(&key);
-            if let RowKey::SecondaryIndex(index_id, _, _) = row_key {
-                if index_id == self.index_id(secondary_id) {
-                    batch.delete(key);
-                } else {
-                    return Ok(());
-                }
-            }
-        }
-        Ok(())
-    }
-
-    fn all_rows(&self) -> Result<Vec<IdRow<Self::T>>, CubeError> {
-        let mut res = Vec::new();
-        for row in self.table_scan(self.snapshot())? {
-            res.push(row?);
-        }
-        Ok(res)
-    }
-
-    fn scan_all_rows<'a>(&'a self) -> Result<TableScanIter<'a, Self>, CubeError> {
-        self.table_scan(self.snapshot())
-    }
-
-    fn scan_rows_by_index<'a, K: Debug>(
-        &'a self,
-        row_key: &'a K,
-        secondary_index: &'a impl RocksSecondaryIndex<Self::T, K>,
-    ) -> Result<IndexScanIter<'a, Self>, CubeError>
-    where
-        K: Hash,
-    {
-        let ref db = self.snapshot();
-
-        let secondary_key_hash = secondary_index
-            .typed_key_hash(&row_key)
-            .to_be_bytes()
-            .to_vec();
-        let secondary_key_val = secondary_index.key_to_bytes(&row_key);
-
-        let key_len = secondary_key_hash.len();
-        let key_min = RowKey::SecondaryIndex(
-            self.index_id(RocksSecondaryIndex::get_id(secondary_index)),
-            secondary_key_hash.clone(),
-            0,
-        );
-
-        let mut opts = ReadOptions::default();
-        opts.set_prefix_same_as_start(true);
-        let iter = db.iterator_opt(
-            IteratorMode::From(&key_min.to_bytes()[0..(key_len + 5)], Direction::Forward),
-            opts,
-        );
-
-        Ok(IndexScanIter {
-            table: self,
-            secondary_key_val,
-            secondary_key_hash,
-            iter,
-        })
-    }
-
-    fn table_scan<'a>(&'a self, db: &'a Snapshot) -> Result<TableScanIter<'a, Self>, CubeError> {
-        let my_table_id = self.table_id();
-        let key_min = RowKey::Table(my_table_id, 0);
-
-        let mut opts = ReadOptions::default();
-        opts.set_prefix_same_as_start(true);
-        let iterator = db.iterator_opt(
-            IteratorMode::From(
-                &key_min.to_bytes()[0..get_fixed_prefix()],
-                Direction::Forward,
-            ),
-            opts,
-        );
-
-        Ok(TableScanIter {
-            table_id: my_table_id,
-            iter: iterator,
-            table: self,
-        })
-    }
-
-    fn build_path_rows<C: Clone, P>(
-        &self,
-        children: Vec<IdRow<C>>,
-        mut parent_id_fn: impl FnMut(&IdRow<C>) -> u64,
-        mut path_fn: impl FnMut(IdRow<C>, Arc<IdRow<Self::T>>) -> P,
-    ) -> Result<Vec<P>, CubeError> {
-        let id_to_child = children
-            .into_iter()
-            .map(|c| (parent_id_fn(&c), c))
-            .collect::<Vec<_>>();
-        let ids = id_to_child
-            .iter()
-            .map(|(id, _)| *id)
-            .unique()
-            .collect::<Vec<_>>();
-        let rows = ids
-            .into_iter()
-            .map(|id| -> Result<(u64, Arc<IdRow<Self::T>>), CubeError> {
-                Ok((id, Arc::new(self.get_row_or_not_found(id)?)))
-            })
-            .collect::<Result<HashMap<_, _>, _>>()?;
-        Ok(id_to_child
-            .into_iter()
-            .map(|(id, c)| path_fn(c, rows.get(&id).unwrap().clone()))
-            .collect::<Vec<_>>())
-    }
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug)]
-pub enum WriteBatchEntry {
-    Put { key: Box<[u8]>, value: Box<[u8]> },
-    Delete { key: Box<[u8]> },
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug)]
-pub struct WriteBatchContainer {
-    entries: Vec<WriteBatchEntry>,
-}
-
-impl WriteBatchContainer {
-    fn new() -> Self {
-        Self {
-            entries: Vec::new(),
-        }
-    }
-
-    fn write_batch(&self) -> WriteBatch {
-        let mut batch = WriteBatch::default();
-        for entry in self.entries.iter() {
-            match entry {
-                WriteBatchEntry::Put { key, value } => batch.put(key, value),
-                WriteBatchEntry::Delete { key } => batch.delete(key),
-            }
-        }
-        batch
-    }
-
-    async fn write_to_file(&self, file_name: &str) -> Result<(), CubeError> {
-        let mut ser = flexbuffers::FlexbufferSerializer::new();
-        self.serialize(&mut ser)?;
-        let mut file = File::create(file_name).await?;
-        Ok(tokio::io::AsyncWriteExt::write_all(&mut file, ser.view()).await?)
-    }
-
-    async fn read_from_file(file_name: &str) -> Result<Self, CubeError> {
-        let mut file = File::open(file_name).await?;
-
-        let mut buffer = Vec::new();
-        tokio::io::AsyncReadExt::read_to_end(&mut file, &mut buffer).await?;
-        let r = flexbuffers::Reader::get_root(&buffer)?;
-        Ok(Self::deserialize(r)?)
-    }
-}
-
-impl WriteBatchIterator for WriteBatchContainer {
-    fn put(&mut self, key: Box<[u8]>, value: Box<[u8]>) {
-        self.entries.push(WriteBatchEntry::Put { key, value });
-    }
-
-    fn delete(&mut self, key: Box<[u8]>) {
-        self.entries.push(WriteBatchEntry::Delete { key });
-    }
 }
 
 fn meta_store_merge(
@@ -2207,9 +1119,7 @@ impl RocksMetaStore {
             let table_ref = DbTableRef {
                 db: &meta_store_to_move.db,
                 snapshot: &meta_store_to_move.db.snapshot(),
-                mem_seq: MemorySequence {
-                    seq_store: meta_store_to_move.seq_store.clone(),
-                },
+                mem_seq: MemorySequence::new(meta_store_to_move.seq_store.clone()),
             };
 
             if let Err(e) = check_indexes_for_all_tables(table_ref) {
@@ -2234,9 +1144,7 @@ impl RocksMetaStore {
         R: Send + Sync + 'static,
     {
         let db = self.db.clone();
-        let mem_seq = MemorySequence {
-            seq_store: self.seq_store.clone(),
-        };
+        let mem_seq = MemorySequence::new(self.seq_store.clone());
         let db_to_send = db.clone();
         let cached_tables = self.cached_tables.clone();
         let rw_loop_sender = self.rw_loop_tx.clone();
@@ -2432,9 +1340,7 @@ impl RocksMetaStore {
         F: for<'a> FnOnce(DbTableRef<'a>) -> Result<R, CubeError> + Send + Sync + 'static,
         R: Send + Sync + 'static,
     {
-        let mem_seq = MemorySequence {
-            seq_store: self.seq_store.clone(),
-        };
+        let mem_seq = MemorySequence::new(self.seq_store.clone());
         let db_to_send = self.db.clone();
 
         let rw_loop_sender = self.rw_loop_tx.clone();
@@ -2474,9 +1380,7 @@ impl RocksMetaStore {
         F: for<'a> FnOnce(DbTableRef<'a>) -> Result<R, CubeError> + Send + Sync + 'static,
         R: Send + Sync + 'static,
     {
-        let mem_seq = MemorySequence {
-            seq_store: self.seq_store.clone(),
-        };
+        let mem_seq = MemorySequence::new(self.seq_store.clone());
         let db_to_send = self.db.clone();
 
         cube_ext::spawn_blocking(move || {
@@ -2678,7 +1582,7 @@ impl RocksMetaStore {
 
         // Validate the columns match types specified in the MultiIndex.
         if let Some(mi) = multi_index {
-            let mi = &mi.row;
+            let mi = mi.get_row();
             if sorted_key_size as usize != mi.key_columns().len() {
                 return Err(CubeError::user(format!(
                     "Partitioned index '{}' has {} columns, got {} columns in ADD TO PARTITIONED INDEX",
@@ -2707,20 +1611,23 @@ impl RocksMetaStore {
             sorted_key_size,
             // Seq column shouldn't participate in partition split. Otherwise we can't do shared nothing calculations across partitions.
             table_id.get_row().seq_column().map(|_| sorted_key_size - 1),
-            multi_index.map(|i| i.id),
+            multi_index.map(|i| i.get_id()),
             IndexType::Regular,
         )?;
         let index_id = rocks_index.insert(index, batch_pipe)?;
         if multi_partitions.is_empty() {
-            rocks_partition.insert(Partition::new(index_id.id, None, None, None), batch_pipe)?;
+            rocks_partition.insert(
+                Partition::new(index_id.get_id(), None, None, None),
+                batch_pipe,
+            )?;
         } else {
             for p in multi_partitions {
                 rocks_partition.insert(
                     Partition::new(
-                        index_id.id,
-                        Some(p.id),
-                        p.row.min_row().cloned(),
-                        p.row.max_row().cloned(),
+                        index_id.get_id(),
+                        Some(p.get_id()),
+                        p.get_row().min_row().cloned(),
+                        p.get_row().max_row().cloned(),
                     ),
                     batch_pipe,
                 )?;
@@ -2807,7 +1714,10 @@ impl RocksMetaStore {
         )?;
 
         let index_id = rocks_index.insert(index, batch_pipe)?;
-        rocks_partition.insert(Partition::new(index_id.id, None, None, None), batch_pipe)?;
+        rocks_partition.insert(
+            Partition::new(index_id.get_id(), None, None, None),
+            batch_pipe,
+        )?;
         Ok(index_id)
     }
 
@@ -3127,7 +2037,7 @@ impl MetaStore for RocksMetaStore {
                                         "Aggregate function {} not allowed for column type {}",
                                         function, &column.column_type
                                         )
-))
+                            ))
                         }
                         Ok(AggregateColumnIndex::new(index, function))
                     })
@@ -5022,13 +3932,12 @@ fn swap_active_partitions_impl(
 
     // Rows are compacted using unique key columns or aggregating index and totals don't match
     let skip_row_count_sanity_check = if let Some(current) = current_active.first() {
-        let current_partition =
-            table
-                .get_row(current.0.id)?
-                .ok_or(CubeError::internal(format!(
-                    "Current partition is not found during swap active: {}",
-                    current.0.id
-                )))?;
+        let current_partition = table
+            .get_row(current.0.get_id())?
+            .ok_or(CubeError::internal(format!(
+                "Current partition is not found during swap active: {}",
+                current.0.get_id()
+            )))?;
         let index = index_table.get_row_or_not_found(current_partition.get_row().get_index_id())?;
         let table = table_table.get_row_or_not_found(index.get_row().table_id())?;
         index.get_row().get_type() == IndexType::Aggregate
@@ -5042,10 +3951,10 @@ fn swap_active_partitions_impl(
 
     let mut dropped_current = HashSet::new();
     for (current_i, (current, chunks)) in current_active.iter().enumerate() {
-        let current_partition = match table.get_row(current.id)? {
+        let current_partition = match table.get_row(current.get_id())? {
             None => {
                 on_dropped_current_partition(current_i)?;
-                dropped_current.insert(current.id);
+                dropped_current.insert(current.get_id());
                 continue;
             }
             Some(p) => p,
@@ -5083,9 +3992,9 @@ fn swap_active_partitions_impl(
 
     for i in 0..new_active.len() {
         let (new, new_file_size) = &new_active[i];
-        if dropped_current.contains(&new.row.parent_partition_id.unwrap()) {
+        if dropped_current.contains(&new.get_row().parent_partition_id.unwrap()) {
             on_dropped_new_partition(i)?;
-            if let Err(e) = table.delete(new.id, batch_pipe) {
+            if let Err(e) = table.delete(new.get_id(), batch_pipe) {
                 // This might happen during DROP TABLE.
                 log::trace!(
                     "Failure when removing new partition, likely not an error: {}",
@@ -5094,10 +4003,12 @@ fn swap_active_partitions_impl(
             }
             continue;
         }
-        let new_partition = table.get_row(new.id)?.ok_or(CubeError::internal(format!(
-            "New partition is not found during swap active: {}",
-            new.id
-        )))?;
+        let new_partition = table
+            .get_row(new.get_id())?
+            .ok_or(CubeError::internal(format!(
+                "New partition is not found during swap active: {}",
+                new.get_id()
+            )))?;
         if new_partition.get_row().is_active() {
             return Err(CubeError::internal(format!(
                 "New partition is already active: {:?}",
@@ -5149,9 +4060,9 @@ fn swap_active_partitions_impl(
             "Deactivated row count ({}) doesn't match activated row count ({}) during swap of partition ({}) and ({}) chunks to new partitions ({})",
             deactivated_row_count,
             activated_row_count,
-            current_active.iter().map(|(p,_)| p.id).join(", "),
-            current_active.iter().flat_map(|(_, cs)| cs).map(|c| c.id).join(", "),
-            new_active.iter().map(|p| p.0.id).join(", ")
+            current_active.iter().map(|(p,_)| p.get_id()).join(", "),
+            current_active.iter().flat_map(|(_, cs)| cs).map(|c| c.get_id()).join(", "),
+            new_active.iter().map(|p| p.0.get_id()).join(", ")
         )));
     }
 
@@ -5165,6 +4076,7 @@ mod tests {
     use crate::config::Config;
     use crate::remotefs::LocalDirRemoteFs;
     use futures_timer::Delay;
+    use rocksdb::IteratorMode;
     use std::thread::sleep;
     use std::time::Duration;
     use std::{env, fs};
@@ -6427,10 +5339,10 @@ impl RocksMetaStore {
                     uploaded_ids_and_sizes.iter().map(|(id, _)| id).join(", ")
                 )));
             }
-            deactivated_row_count += chunk.row.row_count;
+            deactivated_row_count += chunk.get_row().row_count;
             *partition_to_row_diffs
-                .entry(chunk.row.partition_id)
-                .or_default() -= chunk.row.row_count as i64;
+                .entry(chunk.get_row().partition_id)
+                .or_default() -= chunk.get_row().row_count as i64;
             chunks.update_with_fn(*id, |row| row.deactivate(), batch_pipe)?;
         }
         for (id, file_size) in uploaded_ids_and_sizes.iter() {
@@ -6443,10 +5355,10 @@ impl RocksMetaStore {
                     uploaded_ids_and_sizes.iter().map(|(id, _)| id).join(", ")
                 )));
             }
-            activated_row_count += chunk.row.row_count;
+            activated_row_count += chunk.get_row().row_count;
             *partition_to_row_diffs
-                .entry(chunk.row.partition_id)
-                .or_default() += chunk.row.row_count as i64;
+                .entry(chunk.get_row().partition_id)
+                .or_default() += chunk.get_row().row_count as i64;
             chunks.update_with_res_fn(
                 *id,
                 |row| {
@@ -6473,7 +5385,7 @@ impl RocksMetaStore {
         let mut multipart_to_row_diffs = HashMap::<u64, i64>::new();
         for (p, diff) in partition_to_row_diffs {
             let p = partitions.get_row_or_not_found(p)?;
-            let m = match p.row.multi_partition_id {
+            let m = match p.get_row().multi_partition_id {
                 None => continue,
                 Some(m) => m,
             };
@@ -6517,23 +5429,23 @@ impl RocksMetaStore {
         let multi_partitions_table = MultiPartitionRocksTable::new(db.clone());
         for partition in partitions.into_iter() {
             let mut removed_rows = 0;
-            if partition.row.is_active() {
-                removed_rows += partition.row.main_table_row_count;
+            if partition.get_row().is_active() {
+                removed_rows += partition.get_row().main_table_row_count;
             }
             let chunks = chunks_table.get_rows_by_index(
                 &ChunkIndexKey::ByPartitionId(partition.get_id()),
                 &ChunkRocksIndex::PartitionId,
             )?;
             for chunk in chunks.into_iter() {
-                if chunk.row.active {
-                    removed_rows += chunk.row.row_count;
+                if chunk.get_row().active {
+                    removed_rows += chunk.get_row().row_count;
                 }
                 chunks_table.delete(chunk.get_id(), pipe)?;
             }
             partitions_table.delete(partition.get_id(), pipe)?;
 
             if update_multi_partitions {
-                if let Some(m) = partition.row.multi_partition_id {
+                if let Some(m) = partition.get_row().multi_partition_id {
                     multi_partitions_table.update_with_fn(
                         m,
                         |r| r.subtract_rows(removed_rows),
