+++ rust/cubestore/cubestore/src/metastore/rocks_store.rs
@@ -1,0 +1,333 @@
+use crate::metastore::MetaStoreEvent;
+use crate::CubeError;
+use async_trait::async_trait;
+use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
+use rocksdb::{Snapshot, WriteBatch, WriteBatchIterator, DB};
+use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
+use std::fmt::Debug;
+use std::io::Cursor;
+use std::sync::{Arc, Mutex};
+use tokio::fs::File;
+
+macro_rules! enum_from_primitive_impl {
+    ($name:ident, $( $variant:ident )*) => {
+        impl From<u32> for $name {
+            fn from(n: u32) -> Self {
+                $( if n == $name::$variant as u32 {
+                    $name::$variant
+                } else )* {
+                    panic!("Unknown {}: {}", stringify!($name), n);
+                }
+            }
+        }
+    };
+}
+
+macro_rules! enum_from_primitive {
+    (
+        $( #[$enum_attr:meta] )*
+        pub enum $name:ident {
+            $( $( $( #[$variant_attr:meta] )* $variant:ident ),+ = $discriminator:expr ),*
+        }
+    ) => {
+        $( #[$enum_attr] )*
+        pub enum $name {
+            $( $( $( #[$variant_attr] )* $variant ),+ = $discriminator ),*
+        }
+        enum_from_primitive_impl! { $name, $( $( $variant )+ )* }
+    };
+}
+
+enum_from_primitive! {
+    #[derive(Copy, Clone, Eq, PartialEq, Debug, Serialize, Deserialize, Hash)]
+    pub enum TableId {
+        Schemas = 0x0100,
+        Tables = 0x0200,
+        Indexes = 0x0300,
+        Partitions = 0x0400,
+        Chunks = 0x0500,
+        WALs = 0x0600,
+        Jobs = 0x0700,
+        Sources = 0x0800,
+        MultiIndexes = 0x0900,
+        MultiPartitions = 0x0A00
+    }
+}
+
+pub fn get_fixed_prefix() -> usize {
+    13
+}
+
+pub type SecondaryKey = Vec<u8>;
+pub type IndexId = u32;
+
+#[derive(Clone)]
+pub struct MemorySequence {
+    seq_store: Arc<Mutex<HashMap<TableId, u64>>>,
+}
+
+impl MemorySequence {
+    pub fn new(seq_store: Arc<Mutex<HashMap<TableId, u64>>>) -> Self {
+        Self { seq_store }
+    }
+
+    pub fn next_seq(&self, table_id: TableId, snapshot_value: u64) -> Result<u64, CubeError> {
+        let mut store = self.seq_store.lock()?;
+        let mut current = *store.entry(table_id).or_insert(snapshot_value);
+        current += 1;
+        store.insert(table_id, current);
+        Ok(current)
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, Hash, Eq, PartialEq)]
+pub enum RowKey {
+    Table(TableId, u64),
+    Sequence(TableId),
+    SecondaryIndex(IndexId, SecondaryKey, u64),
+    SecondaryIndexInfo { index_id: IndexId },
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, Hash, Eq, PartialEq)]
+pub struct SecondaryIndexInfo {
+    pub version: u32,
+}
+
+impl RowKey {
+    pub fn from_bytes(bytes: &[u8]) -> RowKey {
+        let mut reader = Cursor::new(bytes);
+        match reader.read_u8().unwrap() {
+            1 => RowKey::Table(TableId::from(reader.read_u32::<BigEndian>().unwrap()), {
+                // skip zero for fixed key padding
+                reader.read_u64::<BigEndian>().unwrap();
+                reader.read_u64::<BigEndian>().unwrap()
+            }),
+            2 => RowKey::Sequence(TableId::from(reader.read_u32::<BigEndian>().unwrap())),
+            3 => {
+                let table_id = IndexId::from(reader.read_u32::<BigEndian>().unwrap());
+                let mut secondary_key: SecondaryKey = SecondaryKey::new();
+                let sc_length = bytes.len() - 13;
+                for _i in 0..sc_length {
+                    secondary_key.push(reader.read_u8().unwrap());
+                }
+                let row_id = reader.read_u64::<BigEndian>().unwrap();
+
+                RowKey::SecondaryIndex(table_id, secondary_key, row_id)
+            }
+            4 => {
+                let index_id = IndexId::from(reader.read_u32::<BigEndian>().unwrap());
+
+                RowKey::SecondaryIndexInfo { index_id }
+            }
+            v => panic!("Unknown key prefix: {}", v),
+        }
+    }
+
+    pub fn to_bytes(&self) -> Vec<u8> {
+        let mut wtr = vec![];
+        match self {
+            RowKey::Table(table_id, row_id) => {
+                wtr.write_u8(1).unwrap();
+                wtr.write_u32::<BigEndian>(*table_id as u32).unwrap();
+                wtr.write_u64::<BigEndian>(0).unwrap();
+                wtr.write_u64::<BigEndian>(row_id.clone()).unwrap();
+            }
+            RowKey::Sequence(table_id) => {
+                wtr.write_u8(2).unwrap();
+                wtr.write_u32::<BigEndian>(*table_id as u32).unwrap();
+            }
+            RowKey::SecondaryIndex(index_id, secondary_key, row_id) => {
+                wtr.write_u8(3).unwrap();
+                wtr.write_u32::<BigEndian>(*index_id as IndexId).unwrap();
+                for &n in secondary_key {
+                    wtr.write_u8(n).unwrap();
+                }
+                wtr.write_u64::<BigEndian>(row_id.clone()).unwrap();
+            }
+            RowKey::SecondaryIndexInfo { index_id } => {
+                wtr.write_u8(4).unwrap();
+                wtr.write_u32::<BigEndian>(*index_id as IndexId).unwrap();
+            }
+        }
+        wtr
+    }
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug)]
+pub enum WriteBatchEntry {
+    Put { key: Box<[u8]>, value: Box<[u8]> },
+    Delete { key: Box<[u8]> },
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug)]
+pub struct WriteBatchContainer {
+    entries: Vec<WriteBatchEntry>,
+}
+
+impl WriteBatchContainer {
+    pub fn new() -> Self {
+        Self {
+            entries: Vec::new(),
+        }
+    }
+
+    pub fn write_batch(&self) -> WriteBatch {
+        let mut batch = WriteBatch::default();
+        for entry in self.entries.iter() {
+            match entry {
+                WriteBatchEntry::Put { key, value } => batch.put(key, value),
+                WriteBatchEntry::Delete { key } => batch.delete(key),
+            }
+        }
+        batch
+    }
+
+    pub async fn write_to_file(&self, file_name: &str) -> Result<(), CubeError> {
+        let mut ser = flexbuffers::FlexbufferSerializer::new();
+        self.serialize(&mut ser)?;
+        let mut file = File::create(file_name).await?;
+        Ok(tokio::io::AsyncWriteExt::write_all(&mut file, ser.view()).await?)
+    }
+
+    pub async fn read_from_file(file_name: &str) -> Result<Self, CubeError> {
+        let mut file = File::open(file_name).await?;
+
+        let mut buffer = Vec::new();
+        tokio::io::AsyncReadExt::read_to_end(&mut file, &mut buffer).await?;
+        let r = flexbuffers::Reader::get_root(&buffer)?;
+        Ok(Self::deserialize(r)?)
+    }
+}
+
+impl WriteBatchIterator for WriteBatchContainer {
+    fn put(&mut self, key: Box<[u8]>, value: Box<[u8]>) {
+        self.entries.push(WriteBatchEntry::Put { key, value });
+    }
+
+    fn delete(&mut self, key: Box<[u8]>) {
+        self.entries.push(WriteBatchEntry::Delete { key });
+    }
+}
+
+#[derive(Clone, Serialize, Deserialize, Debug, Eq, PartialEq)]
+pub struct IdRow<T: Clone> {
+    pub(crate) id: u64,
+    pub(crate) row: T,
+}
+
+impl<T: Clone> IdRow<T> {
+    pub fn new(id: u64, row: T) -> IdRow<T> {
+        IdRow { id, row }
+    }
+    pub fn get_id(&self) -> u64 {
+        self.id
+    }
+
+    pub fn get_row(&self) -> &T {
+        &self.row
+    }
+
+    pub fn into_row(self) -> T {
+        self.row
+    }
+}
+
+pub struct KeyVal {
+    pub key: Vec<u8>,
+    pub val: Vec<u8>,
+}
+
+pub struct BatchPipe<'a> {
+    db: &'a DB,
+    write_batch: WriteBatch,
+    events: Vec<MetaStoreEvent>,
+    pub invalidate_tables_cache: bool,
+}
+
+impl<'a> BatchPipe<'a> {
+    pub fn new(db: &'a DB) -> BatchPipe<'a> {
+        BatchPipe {
+            db,
+            write_batch: WriteBatch::default(),
+            events: Vec::new(),
+            invalidate_tables_cache: false,
+        }
+    }
+
+    pub fn batch(&mut self) -> &mut WriteBatch {
+        &mut self.write_batch
+    }
+
+    pub fn add_event(&mut self, event: MetaStoreEvent) {
+        self.events.push(event);
+    }
+
+    pub fn batch_write_rows(self) -> Result<Vec<MetaStoreEvent>, CubeError> {
+        let db = self.db;
+        db.write(self.write_batch)?;
+        Ok(self.events)
+    }
+
+    pub fn invalidate_tables_cache(&mut self) {
+        self.invalidate_tables_cache = true;
+    }
+}
+
+#[derive(Clone)]
+pub struct DbTableRef<'a> {
+    pub db: &'a DB,
+    pub snapshot: &'a Snapshot<'a>,
+    pub mem_seq: MemorySequence,
+}
+
+#[async_trait]
+pub trait MetaStoreTable: Send + Sync {
+    type T: Serialize + Clone + Debug + 'static;
+
+    async fn all_rows(&self) -> Result<Vec<IdRow<Self::T>>, CubeError>;
+
+    async fn row_by_id_or_not_found(&self, id: u64) -> Result<IdRow<Self::T>, CubeError>;
+
+    async fn delete(&self, id: u64) -> Result<IdRow<Self::T>, CubeError>;
+}
+
+#[macro_export]
+macro_rules! meta_store_table_impl {
+    ($name: ident, $table: ty, $rocks_table: ident) => {
+        pub struct $name {
+            rocks_meta_store: RocksMetaStore,
+        }
+
+        impl $name {
+            fn table<'a>(db: DbTableRef<'a>) -> $rocks_table<'a> {
+                <$rocks_table>::new(db)
+            }
+        }
+
+        #[async_trait]
+        impl MetaStoreTable for $name {
+            type T = $table;
+
+            async fn all_rows(&self) -> Result<Vec<IdRow<Self::T>>, CubeError> {
+                self.rocks_meta_store
+                    .read_operation_out_of_queue(move |db_ref| Ok(Self::table(db_ref).all_rows()?))
+                    .await
+            }
+
+            async fn row_by_id_or_not_found(&self, id: u64) -> Result<IdRow<Self::T>, CubeError> {
+                self.rocks_meta_store
+                    .read_operation(move |db_ref| Ok(Self::table(db_ref).get_row_or_not_found(id)?))
+                    .await
+            }
+
+            async fn delete(&self, id: u64) -> Result<IdRow<Self::T>, CubeError> {
+                self.rocks_meta_store
+                    .write_operation(
+                        move |db_ref, batch| Ok(Self::table(db_ref).delete(id, batch)?),
+                    )
+                    .await
+            }
+        }
+    };
+}
